Changes some variable names
