+++ prover/src/utils.rs
@@ -46,7 +46,7 @@ pub fn calculate_matrix_vector_product(matrix_cells: Vec<MatrixCell>, witness: &
 }
 
 /// List of evaluations for eq(r, x) over the boolean hypercube
-pub fn evaluations_over_boolean_hypercube_for_eq(r: Vec<Field256>) -> Vec<Field256> {
+pub fn calculate_evaluations_over_boolean_hypercube_for_eq(r: Vec<Field256>) -> Vec<Field256> {
     let mut ans = vec![Field256::from(1)];
     for x in r {
         let mut left: Vec<Field256> = ans.clone().into_iter().map(|y| {y.mul(Field256::one().sub(x))}).collect();
@@ -102,12 +102,13 @@ pub fn eval_qubic_poly(poly: &Vec<Field256>, point: &Field256) -> Field256 {
     poly[0] + *point * (poly[1] + *point * (poly[2] + *point * poly[3]))
 }
 
-pub fn extract_witness_and_witness_bound (file_path: &str) -> (Vec<Field256>, Vec<Field256>, Vec<Field256>, Vec<Field256>) {
+pub fn extract_witness_and_witness_bound (file_path: &str) -> (Vec<Field256>, Vec<Field256>, Vec<Field256>, Vec<Field256>, usize) {
     let file = File::open(file_path).expect("Failed to open file");
     let r1cs_with_witness: R1CSWithWitness = serde_json::from_reader(file).expect("Failed to parse JSON with Serde");
     let witness = pad_to_power_of_two(stringvec_to_fieldvec(&r1cs_with_witness.witnesses[0]));
     let witness_bound_a = pad_to_power_of_two(calculate_matrix_vector_product(r1cs_with_witness.a, &witness, r1cs_with_witness.num_constraints));
     let witness_bound_b = pad_to_power_of_two(calculate_matrix_vector_product(r1cs_with_witness.b, &witness, r1cs_with_witness.num_constraints));
     let witness_bound_c = pad_to_power_of_two(calculate_matrix_vector_product(r1cs_with_witness.c, &witness, r1cs_with_witness.num_constraints));
-    (witness_bound_a, witness_bound_b, witness_bound_c, witness)
+    let m = next_power_of_two(witness_bound_a.len());
+    (witness_bound_a, witness_bound_b, witness_bound_c, witness, m)
 }
